backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# OpenAI and AI
openai==1.3.6
openai-agents==0.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # PostgreSQL driver (commented for development)
aiosqlite==0.19.0      # SQLite async driver

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Vector Database & RAG
chromadb==0.4.17
sentence-transformers==2.2.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
pre-commit==3.5.0

# Production
gunicorn==21.2.0
redis==5.0.1